% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_modelviz.R
\name{modelviz_palette}
\alias{modelviz_palette}
\title{The modelviz colour palette}
\usage{
modelviz_palette(n = NULL)
}
\arguments{
\item{n}{Number of colours to return. If \code{NULL} (the default) the full
palette is returned. When \code{n} is larger than the palette,
\code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}} is used to interpolate.}
}
\value{
A character vector of hex colour codes.
}
\description{
A qualitative palette built around the blue/red pairing used for
Bayesian/frequentist comparisons, extended with complementary hues so that
it also works for several groups.
}
\examples{
modelviz_palette(2)
scales::show_col(modelviz_palette())
}
