% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_modelviz.R
\name{scale_colour_modelviz}
\alias{scale_colour_modelviz}
\alias{scale_color_modelviz}
\alias{scale_fill_modelviz}
\title{modelviz colour and fill scales}
\usage{
scale_colour_modelviz(...)

scale_color_modelviz(...)

scale_fill_modelviz(...)
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\value{
A ggplot2 scale that can be added to a plot.
}
\description{
Discrete ggplot2 scales using \code{\link[=modelviz_palette]{modelviz_palette()}}.
}
\examples{
library(ggplot2)
ggplot(crop_yield, aes(rainfall, yield, colour = treatment)) +
  geom_point() +
  scale_colour_modelviz() +
  theme_modelviz()
}
