% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_statviz.R
\name{scale_colour_statviz}
\alias{scale_colour_statviz}
\alias{scale_color_statviz}
\alias{scale_fill_statviz}
\title{statviz colour and fill scales}
\usage{
scale_colour_statviz(...)

scale_color_statviz(...)

scale_fill_statviz(...)
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\value{
A ggplot2 scale that can be added to a plot.
}
\description{
Discrete ggplot2 scales using \code{\link[=statviz_palette]{statviz_palette()}}.
}
\examples{
library(ggplot2)
ggplot(crop_yield, aes(rainfall, yield, colour = treatment)) +
  geom_point() +
  scale_colour_statviz() +
  theme_statviz()
}
