% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{palette_preview}
\alias{palette_preview}
\title{Preview the statviz palette}
\usage{
palette_preview(n = 8)
}
\arguments{
\item{n}{Number of colours to preview.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Displays the colours returned by \code{\link[=statviz_palette]{statviz_palette()}} as labelled swatches --
handy when choosing how many groups to show or when documenting a figure.
}
\examples{
palette_preview()
palette_preview(4)
}
