% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_depictr.R
\name{scale_colour_depictr}
\alias{scale_colour_depictr}
\alias{scale_color_depictr}
\alias{scale_fill_depictr}
\title{depictr colour and fill scales}
\usage{
scale_colour_depictr(...)

scale_color_depictr(...)

scale_fill_depictr(...)
}
\arguments{
\item{...}{Passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\value{
A ggplot2 scale that can be added to a plot.
}
\description{
Discrete ggplot2 scales using \code{\link[=depictr_palette]{depictr_palette()}}.
}
\examples{
library(ggplot2)
ggplot(crop_yield, aes(rainfall, yield, colour = treatment)) +
  geom_point() +
  scale_colour_depictr() +
  theme_depictr()
}
